fix(WebSocketShard): Properly handle RESUMED timeout
Makes RESUMED timeout delay itself if there were events received in the last 20 seconds

+++ packages/discord.js/src/client/websocket/WebSocketShard.js
@@ -191,6 +191,14 @@ class WebSocketShard extends EventEmitter {
      * @private
      */
     Object.defineProperty(this, 'connectedAt', { value: 0, writable: true });
+
+    /**
+     * Time when the last replayed event was received
+     * @name WebSocketShard#lastReplayedAt
+     * @type {number}
+     * @private
+     */
+    Object.defineProperty(this, 'lastReplayedAt', { value: 0, writable: true });
   }
 
   /**
@@ -454,6 +462,10 @@ class WebSocketShard extends EventEmitter {
         this.sendHeartbeat('ResumeHeartbeat');
         break;
       }
+      default: {
+        if (this.status === Status.Resuming) this.lastReplayedAt = Date.now();
+        break;
+      }
     }
 
     if (packet.s > this.sequence) this.sequence = packet.s;
@@ -580,7 +592,7 @@ class WebSocketShard extends EventEmitter {
     this.helloTimeout = setTimeout(() => {
       this.debug('Did not receive HELLO in time. Destroying and connecting again.');
       this.destroy({ reset: true, closeCode: 4009 });
-    }, 20_000);
+    }, 20_000).unref();
   }
 
   /**
@@ -597,11 +609,15 @@ class WebSocketShard extends EventEmitter {
       }
       return;
     }
-    this.debug('Setting a RESUMED dispatch timeout for 30s.');
+    this.debug('Setting a RESUMED dispatch timeout for 20s.');
     this.resumedDispatchTimeout = setTimeout(() => {
+      if ((Date.now() - this.lastReplayedAt) < 20_000) {
+        this.debug('Received a message within the last 20s. Delaying RESUMED timeout.');
+        return this.setResumedDispatchTimeout();
+      }
       this.debug('Did not receive RESUMED in time. Destroying and connecting again.');
       this.destroy({ reset: false, closeCode: 4009 });
-    }, 30_000);
+    }, 20_000).unref();
   }
 
   /**
@@ -622,7 +638,7 @@ class WebSocketShard extends EventEmitter {
     this.readyDispatchTimeout = setTimeout(() => {
       this.debug('Did not receive READY in time. Destroying and connecting again.');
       this.destroy({ reset: true, closeCode: 4009 });
-    }, 20_000);
+    }, 20_000).unref();
   }
 
   /**
